// api/coreService/types/dto/SurahResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type DtoSurahResponse = {
    /**
     * @type integer | undefined
    */
    ayahCount?: number;
    /**
     * @type string | undefined
    */
    createdAt?: string;
    /**
     * @type string | undefined
    */
    createdBy?: string;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type string | undefined
    */
    nameArabic?: string;
    /**
     * @type string | undefined
    */
    nameEnglish?: string;
    /**
     * @type integer | undefined
    */
    revelationOrder?: number;
    /**
     * @type string | undefined
    */
    revelationPlace?: string;
    /**
     * @type string | undefined
    */
    updatedAt?: string;
    /**
     * @type string | undefined
    */
    updatedBy?: string;
};